requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
futures==3.2.0
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-language==1.0.0
googleapis-common-protos==1.5.3
grpcio==1.10.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
